sentry: add sentry to the project (#157)
This PR adds Sentry to the VS Code extension project. It also attempts to fix the pre-release/release publishing issue.

The purpose of integrating Sentry is to capture uncaught errors or specific errors, such as errors during the installation of the LSP server or while running the LSP client. It is not intended to capture incorrectly written queries or connection issues.

This PR does not implement Sentry in the React scripts yet. React exceptions will be addressed in a future PR once this initial one is working properly.